services/core_service/requirements.txt
django==4.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
psycopg2-binary==2.9.9
django-cors-headers==4.3.1
drf-spectacular==0.27.1